+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Develop
  - Added the option to directly  apply filters when generating partly correlated noise
+ - Added the audiotools.cmplx_corr function which calculates the
+   complex-valued correlation coefficent
 
 0.61 -> 0.62
  - Fixed the shape of the channels after appling a filterbank